+++ common/mempool/mempool_ll.h
@@ -22,19 +22,10 @@
  * KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations
  * under the License.
- *
  */
 /*
  * NOTICE: File has been changed from original implementation.
  */
-
-/**
- * @addtogroup OSKernel
- * @{
- *   @defgroup OSMempool Memory Pools
- *   @{
- */
-
 
 #ifndef _OS_MEMPOOL_H_
 #define _OS_MEMPOOL_H_
@@ -51,20 +42,7 @@
 extern "C" {
 #endif
 
-#define MYNEWT_VAL_OS_MEMPOOL_POISON (0)
-#define MYNEWT_VAL_OS_MEMPOOL_CHECK (0)
-
 #define MYNEWT_VAL(_name)                       MYNEWT_VAL_ ## _name
-#define MYNEWT_VAL_CHOICE(_name, _val)          MYNEWT_VAL_ ## _name ## __ ## _val
-
-
-#ifndef min
-#define min(a, b) ((a)<(b)?(a):(b))
-#endif
-
-#ifndef max
-#define max(a, b) ((a)>(b)?(a):(b))
-#endif
 
 #define OS_ALIGN(__n, __a) (                             \
         (((__n) & ((__a) - 1)) == 0)                   ? \
@@ -72,14 +50,6 @@ extern "C" {
             ((__n) + ((__a) - ((__n) & ((__a) - 1))))    \
         )
 #define OS_ALIGNMENT 4
-
-extern SemaphoreHandle_t hosted_port_mutex;
-
-typedef uint32_t os_sr_t;
-
-#define OS_INIT_CRITICAL() (hosted_port_mutex = xSemaphoreCreateMutex())
-#define OS_ENTER_CRITICAL() (xSemaphoreTake((hosted_port_mutex), portMAX_DELAY))
-#define OS_EXIT_CRITICAL() (xSemaphoreGive((hosted_port_mutex)))
 
 enum os_error {
     OS_OK = 0,
@@ -106,8 +76,14 @@ typedef enum os_error os_error_t;
  * caller.
  */
 struct os_memblock {
+    /** Next memory block in the list. */
     SLIST_ENTRY(os_memblock) mb_next;
 };
+
+/* XXX: Change this structure so that we keep the first address in the pool? */
+/* XXX: add memory debug structure and associated code */
+/* XXX: Change how I coded the SLIST_HEAD here. It should be named:
+   SLIST_HEAD(,os_memblock) mp_head; */
 
 /**
  * Memory pool
@@ -125,12 +101,13 @@ struct os_mempool {
     uint8_t mp_flags;
     /** Address of memory buffer used by pool */
     uintptr_t mp_membuf_addr;
+    /** Next memory pool in the list. */
     STAILQ_ENTRY(os_mempool) mp_list;
+    /** Head of the list of memory blocks. */
     SLIST_HEAD(,os_memblock);
     /** Name for memory block */
-    const char *name;
+    char *name;
 };
-
 
 /**
  * Indicates an extended mempool.  Address can be safely cast to
@@ -161,14 +138,19 @@ struct os_mempool_ext;
 typedef os_error_t os_mempool_put_fn(struct os_mempool_ext *ome, void *data,
                                      void *arg);
 
+/** Extended memory pool. */
 struct os_mempool_ext {
+    /** Standard memory pool. */
     struct os_mempool mpe_mp;
 
-    /* Callback that is executed immediately when a block is freed. */
+    /** Callback that is executed immediately when a block is freed. */
     os_mempool_put_fn *mpe_put_cb;
+
+    /** Optional argument passed to the callback function. */
     void *mpe_put_arg;
 };
 
+/** Length of the name of memory pool */
 #define OS_MEMPOOL_INFO_NAME_LEN (32)
 
 /**
@@ -193,21 +175,33 @@ struct os_mempool_info {
  * is NOT in bytes! The size is the number of os_membuf_t elements required for
  * the memory pool.
  */
-#if (OS_CFG_ALIGNMENT == OS_CFG_ALIGN_4)
-#define OS_MEMPOOL_SIZE(n,blksize)      ((((blksize) + 3) / 4) * (n))
-typedef uint32_t os_membuf_t;
+#if MYNEWT_VAL(OS_MEMPOOL_GUARD)
+/** Leave extra 4 bytes of guard area at the end. */
+#define OS_MEMPOOL_BLOCK_SZ(sz) ((sz) + sizeof(os_membuf_t))
 #else
-#define OS_MEMPOOL_SIZE(n,blksize)      ((((blksize) + 7) / 8) * (n))
-typedef uint64_t os_membuf_t;
+/** Size of a memory pool block. */
+#define OS_MEMPOOL_BLOCK_SZ(sz) (sz)
 #endif
+#if (OS_ALIGNMENT == 4)
+typedef uint32_t os_membuf_t;
+#elif (OS_ALIGNMENT == 8)
+typedef uint64_t os_membuf_t;
+#elif (OS_ALIGNMENT == 16)
+typedef __uint128_t os_membuf_t;
+#else
+#error "Unhandled `OS_ALIGNMENT` for `os_membuf_t`"
+#endif /* OS_ALIGNMENT == * */
+
+/** The total size of a memory pool, including alignment. */
+#define OS_MEMPOOL_SIZE(n,blksize)      (((OS_MEMPOOL_BLOCK_SZ(blksize) + ((OS_ALIGNMENT)-1)) / (OS_ALIGNMENT)) * (n))
 
 /** Calculates the number of bytes required to initialize a memory pool. */
 #define OS_MEMPOOL_BYTES(n,blksize)     \
     (sizeof (os_membuf_t) * OS_MEMPOOL_SIZE((n), (blksize)))
 
 struct mempool_ops_t {
-	os_error_t (*mempool_init)(struct os_mempool *mp, uint16_t blocks, uint32_t block_size, void *membuf, const char *name);
-	void       (*mempool_unregister)(struct os_mempool *mp);
+	os_error_t (*mempool_init)(struct os_mempool *mp, uint16_t blocks, uint32_t block_size, void *membuf, char *name);
+	os_error_t (*mempool_unregister)(struct os_mempool *mp);
 	void *     (*memblock_get)(struct os_mempool *mp);
 	os_error_t (*memblock_put)(struct os_mempool *mp, void *block_addr);
 };
@@ -221,9 +215,3 @@ struct mempool_ops_t * os_mempool_get_ops(void);
 #endif // CONFIG_ESP_HOSTED_USE_MEMPOOL
 
 #endif  /* _OS_MEMPOOL_H_ */
-
-
-/**
- *   @} OSMempool
- * @} OSKernel
- */